refactor: move axes components to own folder
